// frontendSample/packages/auth/types/enums.ts
export enum AuthFormScreen {
  SignIn = "sign-in",
  // SignInSuccess = "login-success", // Disabled, redirecting home
  SignUp = "sign-up",
  SignUpSuccess = "sign-up-success",
  ForgotPassword = "forgot-password",
  VerifyResetPasscode = "verify-reset-passcode",
  ResetPassword = "reset-password",
  ResetPasswordSuccess = "reset-password-success",
}